Move HttpApiProblemStdError into dedicated module
Since the type is now used in to different routes, it should really be
it's own module.
