+++ application/comit_node/src/http_api/mod.rs
@@ -7,6 +7,10 @@ pub mod ledger;
 #[macro_use]
 pub mod asset;
 
+mod problem;
+
+pub use self::problem::*;
+
 pub const PATH: &str = "swaps";
 
 mod ledger_impls {
