+++ application/comit_node/src/http_api/rfc003/swap.rs
@@ -4,10 +4,10 @@ use ethereum_support::{self, EtherQuantity};
 use event_store::{self, EventStore};
 use frunk;
 use futures::sync::mpsc::UnboundedSender;
-use http_api;
-use http_api_problem::{HttpApiProblem, HttpStatusCode};
-use hyper::{header, StatusCode};
-use std::{error::Error as StdError, fmt, sync::Arc};
+use http_api::{self, problem::HttpApiProblemStdError};
+use http_api_problem::HttpApiProblem;
+use hyper::header;
+use std::sync::Arc;
 use swap_protocols::{
     asset::Asset,
     ledger::{Bitcoin, Ethereum},
@@ -30,29 +30,6 @@ pub enum Error {
     ClientFactory(comit_client::ClientFactoryError),
     Unsupported,
     NotFound,
-}
-
-#[derive(Debug)]
-pub struct HttpApiProblemStdError {
-    pub inner: HttpApiProblem,
-}
-
-impl From<HttpApiProblem> for HttpApiProblemStdError {
-    fn from(problem: HttpApiProblem) -> Self {
-        Self { inner: problem }
-    }
-}
-
-impl fmt::Display for HttpApiProblemStdError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.inner.title)
-    }
-}
-
-impl StdError for HttpApiProblemStdError {
-    fn source(&self) -> Option<&(dyn StdError + 'static)> {
-        None
-    }
 }
 
 impl From<Error> for HttpApiProblem {
@@ -112,21 +89,6 @@ pub struct SwapCreated {
     pub id: SwapId,
 }
 
-pub fn customize_error(rejection: Rejection) -> Result<impl Reply, Rejection> {
-    if let Some(ref err) = rejection.find_cause::<HttpApiProblemStdError>() {
-        let code = err
-            .inner
-            .status
-            .unwrap_or(HttpStatusCode::InternalServerError);
-        let json = warp::reply::json(&err.inner);
-        return Ok(warp::reply::with_status(
-            json,
-            StatusCode::from_u16(code.to_u16()).unwrap(),
-        ));
-    }
-    Err(rejection)
-}
-
 #[allow(clippy::needless_pass_by_value)]
 pub fn post_swap(
     request_body_kind: SwapRequestBodyKind,
@@ -148,9 +110,9 @@ pub fn post_swap(
             e.into_inner(),
             id
         );
-        return Err(warp::reject::custom(HttpApiProblemStdError {
-            inner: HttpApiProblem::with_title_from_status(500),
-        }));
+        return Err(warp::reject::custom(HttpApiProblemStdError::from(
+            HttpApiProblem::with_title_from_status(500),
+        )));
     }
 
     let swap_created = SwapCreated { id };
@@ -198,9 +160,9 @@ pub fn get_swap<E: EventStore<SwapId>, T: MetadataStore<SwapId>, S: StateStore<S
 
     match result {
         Some(swap_status) => Ok(warp::reply::json(&swap_status)),
-        None => Err(warp::reject::custom(HttpApiProblemStdError {
-            inner: Error::NotFound.into(),
-        })),
+        None => Err(warp::reject::custom(HttpApiProblemStdError::new(
+            Error::NotFound,
+        ))),
     }
 }
