+++ application/comit_node/src/lib.rs
@@ -39,6 +39,7 @@ extern crate event_store;
 extern crate fern;
 extern crate frunk_core;
 extern crate hex;
+extern crate http;
 extern crate http_api_problem;
 extern crate hyper;
 extern crate rand;
